BLOCKCHAIN-523 Start separating proposal components
