+++ src/components/Proposal/Preimage/index.js
@@ -1,0 +1,54 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { useDispatch, useSelector } from 'react-redux';
+import { blockchainActions } from '../../../redux/actions';
+import { blockchainSelectors } from '../../../redux/selectors';
+import { decodeCall } from '../../../api/nodeRpcCall';
+import { Proposal } from '..';
+
+export function Preimage({ hash, len, isDetailsHidden }) {
+  const dispatch = useDispatch();
+  const [call, setCall] = React.useState(null);
+  const preimages = useSelector(blockchainSelectors.preimages);
+  const preimage = preimages[hash.toString()];
+
+  React.useEffect(() => {
+    if (!preimage) {
+      dispatch(blockchainActions.fetchPreimage.call({
+        hash,
+        len,
+      }));
+    }
+  }, [dispatch, preimage, hash, len]);
+
+  React.useEffect(() => {
+    (async () => {
+      if (preimage && preimage.isSome && !call) {
+        setCall(await decodeCall(preimage.unwrap()));
+      }
+    })();
+  }, [preimage, call, setCall]);
+
+  if (preimage === undefined) return <div>Loading details...</div>;
+  if (preimage.isNone) {
+    return (
+      <div>
+        Details not provided on-chain. Hash:
+        {hash.toString()}
+      </div>
+    );
+  }
+  if (call === null) return <div>Loading details...</div>;
+
+  return <Proposal proposal={call} isDetailsHidden={isDetailsHidden} />;
+}
+
+Preimage.propTypes = {
+  // eslint-disable-next-line react/forbid-prop-types
+  hash: PropTypes.object.isRequired,
+  // eslint-disable-next-line react/forbid-prop-types
+  len: PropTypes.object.isRequired,
+  isDetailsHidden: PropTypes.bool.isRequired,
+};
+
+export default Preimage;
