+++ src/components/Proposal/index.js
@@ -1,266 +1,13 @@
-import React, { Fragment, useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import cx from 'classnames';
-import router from '../../router';
-import { blockchainActions, walletActions } from '../../redux/actions';
-import { blockchainSelectors, identitySelectors, walletSelectors } from '../../redux/selectors';
-import { decodeCall } from '../../api/nodeRpcCall';
-import styles from './styles.module.scss';
-import stylesAnim from '../Voting/Referendum/Items/item.module.scss';
+import { walletActions } from '../../redux/actions';
+import { identitySelectors, walletSelectors } from '../../redux/selectors';
 import { formatAssets, formatDollars, formatMerits } from '../../utils/walletHelpers';
-import { useAddIdToContext } from './useAddIdToContect.js';
+import { useAddIdToContext } from './hooks/useAddIdToContext';
 import useRemark from '../../hooks/useRemark';
 /* eslint-disable react/forbid-prop-types */
 const ProposalProp = PropTypes.object;
-
-function Raw({ proposal }) {
-  return <pre className={styles.legislationContent}>{JSON.stringify(proposal.toHuman(), null, 2)}</pre>;
-}
-
-Raw.propTypes = { proposal: ProposalProp.isRequired };
-
-function RepealLegislationSection({ proposal }) {
-  const { args: [tier, { year, index }, section] } = proposal;
-  return (
-    <div>
-      Repeal legislation
-      {' '}
-      <Link to={`${router.home.legislation}/${tier.toString()}`}>
-        {tier.toString()}
-        {' '}
-        -
-        {' '}
-        {year.toNumber()}
-        /
-        {index.toNumber()}
-        {' '}
-        - Section #
-        {section.toNumber()}
-      </Link>
-    </div>
-  );
-}
-RepealLegislationSection.propTypes = { proposal: ProposalProp.isRequired };
-
-function AddLegislation({ proposal, isDetailsHidden }) {
-  const { args: [tier, { year, index }, sections] } = proposal;
-  return (
-    <div className={styles.text}>
-      <p>
-        Add new legislation
-        {' '}
-        <Link to={`${router.home.legislation}/${tier.toString()}`} className={styles.blue}>
-          {tier.toString()}
-        </Link>
-        {' '}
-        -
-        {' '}
-        {year.toNumber()}
-        /
-        {index.toNumber()}
-        {!isDetailsHidden ? '.' : '...'}
-      </p>
-      <div className={cx(stylesAnim.anim, !isDetailsHidden ? stylesAnim.shown : stylesAnim.hidden)}>
-        {sections.map((section, idx) => (
-        // eslint-disable-next-line react/no-array-index-key
-          <Fragment key={idx}>
-            <p>
-              Section #
-              {idx}
-            </p>
-            <p className={styles.legislationContent}>{new TextDecoder('utf-8').decode(section)}</p>
-          </Fragment>
-        ))}
-      </div>
-
-    </div>
-  );
-}
-
-AddLegislation.defaultProps = {
-  isDetailsHidden: false,
-};
-
-AddLegislation.propTypes = {
-  proposal: ProposalProp.isRequired,
-  isDetailsHidden: PropTypes.bool,
-};
-
-function AmendLegislation({ proposal }) {
-  const { args: [tier, { year, index }, section, newContent] } = proposal;
-  return (
-    <div>
-      <p>
-        Amend or add legislation section
-        {' '}
-        <Link to={`${router.home.legislation}/${tier.toString()}`}>
-          {tier.toString()}
-          {' '}
-          -
-          {' '}
-          {year.toNumber()}
-          /
-          {index.toNumber()}
-          {' '}
-          - Section #
-          {section.toNumber()}
-        </Link>
-        .
-      </p>
-      <p>New content:</p>
-      <p className={styles.legislationContent}>{newContent.toHuman()}</p>
-    </div>
-  );
-}
-AmendLegislation.propTypes = { proposal: ProposalProp.isRequired };
-
-function RepealLegislation({ proposal }) {
-  const { args: [tier, { year, index }] } = proposal;
-  return (
-    <div>
-      Repeal legislation
-      {' '}
-      <Link to={`${router.home.legislation}/${tier.toString()}`}>
-        {tier.toString()}
-        {' '}
-        -
-        {' '}
-        {year.toNumber()}
-        /
-        {index.toNumber()}
-      </Link>
-    </div>
-  );
-}
-RepealLegislation.propTypes = { proposal: ProposalProp.isRequired };
-
-export function Preimage({ hash, len, isDetailsHidden }) {
-  const dispatch = useDispatch();
-  const [call, setCall] = useState(null);
-  const preimages = useSelector(blockchainSelectors.preimages);
-  const preimage = preimages[hash.toString()];
-
-  useEffect(() => {
-    if (!preimage) {
-      dispatch(blockchainActions.fetchPreimage.call({
-        hash,
-        len,
-      }));
-    }
-  }, [dispatch, preimage, hash, len]);
-
-  useEffect(() => {
-    (async () => {
-      if (preimage && preimage.isSome && !call) {
-        setCall(await decodeCall(preimage.unwrap()));
-      }
-    })();
-  }, [preimage, call, setCall]);
-
-  if (preimage === undefined) return <div>Loading details...</div>;
-  if (preimage.isNone) {
-    return (
-      <div>
-        Details not provided on-chain. Hash:
-        {hash.toString()}
-      </div>
-    );
-  }
-  if (call === null) return <div>Loading details...</div>;
-
-  return <Proposal proposal={call} isDetailsHidden={isDetailsHidden} />;
-}
-
-Preimage.propTypes = {
-  hash: PropTypes.object.isRequired,
-  len: PropTypes.object.isRequired,
-  isDetailsHidden: PropTypes.bool.isRequired,
-};
-
-function FastTrackedReferendum({ proposal, fastTrack }) {
-  const { hash_: hash, len } = proposal.args[0].asLookup;
-  const [, votingPeriodInBlocks, enactmentPeriodInBlocks] = fastTrack.args;
-  const votingPeriodInDays = (votingPeriodInBlocks.toNumber() * 6) / 3600 / 24;
-  const enactmentPeriodInDays = (enactmentPeriodInBlocks.toNumber() * 6) / 3600 / 24;
-  return (
-    <div>
-      Immediately start new referendum with shortened Referendum and/or Enactment Period.
-      <ul>
-        <li>
-          Referendum Period:
-          {votingPeriodInDays}
-          {' '}
-          day(s)
-        </li>
-        <li>
-          Enactment Period:
-          {enactmentPeriodInDays}
-          {' '}
-          day(s)
-        </li>
-        <li>
-          <Preimage {...{ hash, len }} />
-        </li>
-      </ul>
-    </div>
-  );
-}
-FastTrackedReferendum.propTypes = { proposal: ProposalProp.isRequired, fastTrack: ProposalProp.isRequired };
-
-function Referendum({ proposal }) {
-  const { hash_: hash, len } = proposal.args[0].asLookup;
-  return (
-    <div>
-      Propose a new referendum with simplified tally rules (simple majority of votes).
-      <Preimage {...{ hash, len }} />
-    </div>
-  );
-}
-Referendum.propTypes = { proposal: ProposalProp.isRequired };
-
-function BatchAll({ proposal }) {
-  function fastTrackMatches(prop, fastTrack) {
-    const fastTrackHash = fastTrack.args[0];
-    const p = prop.args[0];
-    if (p.isLookup) return p.asLookup.hash_.eq(fastTrackHash);
-
-    // our FE only uses Lookup
-    return false;
-  }
-
-  const { args: [calls] } = proposal;
-  if (calls.length === 2
-    && calls[0].section === 'democracy'
-    && calls[0].method === 'externalPropose'
-    && calls[1].section === 'democracy'
-    && calls[1].method === 'fastTrack'
-    && fastTrackMatches(calls[0], calls[1])) {
-    return <FastTrackedReferendum proposal={calls[0]} fastTrack={calls[1]} />;
-  }
-  return calls.map((call, idx) => (
-    // eslint-disable-next-line react/no-array-index-key
-    <div key={`proposal ${idx}`}>
-      <Proposal proposal={call} />
-      <br />
-    </div>
-  ));
-}
-BatchAll.propTypes = { proposal: ProposalProp.isRequired };
-
-function Blacklist({ proposal }) {
-  const hash = proposal.args[0];
-  return (
-    <div>
-      Blacklist and cancel referendum proposal.
-      {' '}
-      { proposal.args[1].isSome && `Cancel ongoing referendum #${proposal.args[1].unwrap().toString()}.`}
-      <Preimage hash={hash} len={null} />
-    </div>
-  );
-}
-Blacklist.propTypes = { proposal: ProposalProp.isRequired };
 
 function CouncilSenateExecute({ proposal }) {
   const { args: [calls] } = proposal;
